+++ source/engines/referenceEngine.js
@@ -41,7 +41,7 @@ function removeClub(id, channelManager) {
 	updateListReference(channelManager, "club");
 }
 
-/** @type {{petition: {channelId: string; messageId: string}, club: {channelId: string; messageId: string;}, rules: {channelId: string; messageId: string;}, "press-kit": {channelId: string; messageId: string;}, "proxy-thread-info": {channelId: string; messageId: string;}}} */
+/** @type {{ petition: { channelId: string; messageId: string; }; club: { channelId: string; messageId: string; }; rules: { channelId: string; messageId: string; }; "press-kit": { channelId: string; messageId: string; } }} */
 let referenceMessages = require('../../config/referenceMessageIds.json');
 
 /** Builds the MessageOptions for the petition list message